feat: allow to specify all sentry options
This also allows the client to specify beforeSend, while still
having the throttle functionality.

BREAKING CHANGE: The release, environment and sentryDns attributes
to initSentry should now be specified inside the options attribute.
The buildTimestamp attribute is renamed to buildTime to follow what
we actually store as tag name:

```ts
initSentry({
  options: {
    release: "xxx",
    environment: "prod",
    dns: "xxx",
  },
  buildTime: "xxx",
})
```