chore: rework Jenkinsfile and build step
This changes so that the build is done during the "prepare"
step, allowing us to depend on the library by a Git commit.
This also means the build is happening at the time of install,
so the order is a bit changed.

The semantic release setup is also simplified because of this.
See related docs at https://confluence.capraconsulting.no/x/x6QBC