feat: rename log helpers
This is an attempt to make the code more understandable when logging
events. By just having method names such as "error", it was difficult
to understand that this would capture an event and not trigger an error.

Using capture in the name follows the convention that Sentry uses
itself.

BREAKING CHANGE: The log helpers are now named e.g. captureError instead
of error.